refactor: move writeCommand and CachedFP struct to another directory
