+++ executor/cache.go
@@ -1,11 +1,10 @@
 package executor
 
 import (
-	"fmt"
 	"sync/atomic"
 	"time"
 
-	"github.com/alpacahq/marketstore/v4/utils/io"
+	"github.com/alpacahq/marketstore/v4/executor/wal"
 )
 
 /*
@@ -15,26 +14,12 @@ import (
 */
 const WriteChannelCommandDepth = 1000000
 
-type WriteCommand struct {
-	RecordType    io.EnumRecordType
-	WALKeyPath    string
-	VarRecLen     int
-	Offset, Index int64
-	Data          []byte
-	DataShapes    []io.DataShape
-}
-
-// Convert WriteCommand to string for debuging/presentation
-func (wc *WriteCommand) toString() string {
-	return fmt.Sprintf("WC[%v] WALKeyPath:%s (len:%d, off:%d, idx:%d, dsize:%d)", wc.RecordType, wc.WALKeyPath, wc.VarRecLen, wc.Offset, wc.Index, len(wc.Data))
-}
-
 // TransactionPipe stores the contents of the current pending Transaction Group
 // and writes it to WAL when flush() is called
 type TransactionPipe struct {
-	tgID         int64              // Current transaction group ID
-	writeChannel chan *WriteCommand // Channel for write commands
-	flushChannel chan chan struct{} // Channel for flush request
+	tgID         int64                  // Current transaction group ID
+	writeChannel chan *wal.WriteCommand // Channel for write commands
+	flushChannel chan chan struct{}     // Channel for flush request
 }
 
 // NewTransactionPipe creates a new transaction pipe that channels all
@@ -42,7 +27,7 @@ type TransactionPipe struct {
 func NewTransactionPipe() *TransactionPipe {
 	tgc := new(TransactionPipe)
 	// Allocate the write channel with enough depth to allow all conceivable writers concurrent access
-	tgc.writeChannel = make(chan *WriteCommand, WriteChannelCommandDepth)
+	tgc.writeChannel = make(chan *wal.WriteCommand, WriteChannelCommandDepth)
 	tgc.flushChannel = make(chan chan struct{}, WriteChannelCommandDepth)
 	tgc.NewTGID()
 	return tgc
